+++ src/data_gradients/batch_processors/adapters/tensor_extractor.py
