+++ src/data_gradients/batch_processors/segmentation.py
@@ -1,7 +1,7 @@
 from typing import List, Optional, Callable
 
 from data_gradients.batch_processors.base import BatchProcessor
-from data_gradients.batch_processors.extractors.batch_extractor import BatchExtractor
+from data_gradients.batch_processors.adapters.dataset_adapter import DatasetAdapter
 from data_gradients.batch_processors.preprocessors.segmentation import SegmentationBatchPreprocessor
 from data_gradients.batch_processors.formatters.segmentation import SegmentationBatchFormatter
 
@@ -17,7 +17,7 @@ class SegmentationBatchProcessor(BatchProcessor):
         images_extractor: Optional[Callable] = None,
         labels_extractor: Optional[Callable] = None,
     ):
-        extractor = BatchExtractor(
+        dataset_adapter = DatasetAdapter(
             images_extractor=images_extractor,
             labels_extractor=labels_extractor,
         )
@@ -29,4 +29,4 @@ class SegmentationBatchProcessor(BatchProcessor):
         )
         preprocessor = SegmentationBatchPreprocessor()
 
-        super().__init__(batch_extractor=extractor, batch_formatter=formatter, batch_preprocessor=preprocessor)
+        super().__init__(dataset_adapter=dataset_adapter, batch_formatter=formatter, batch_preprocessor=preprocessor)
